CB-261 refactor: modify mypage layout
마이페이지 레이아웃 구성 수정
- 최상단에서 관리하던 것을 각 페이지 파일에서 레이아웃을 적용하도록 수정
- 동적 라우팅에 따른 관리 어려움으로 인해 변경
