+++ src/pages/Mypage/Profile.tsx
@@ -1,7 +1,11 @@
+import Layout from '../../components/layout/Layout';
+
 export default function ProfilePage() {
   return (
-    <div className="w-full">
-      <h3>프로필 수정 페이지</h3>
-    </div>
+    <Layout header title="프로필 수정" canGoBack>
+      <div className="w-full">
+        <h3>프로필 수정 페이지</h3>
+      </div>
+    </Layout>
   );
 }
