+++ src/pages/Mypage/Wish.tsx
@@ -1,7 +1,11 @@
+import Layout from '@/components/layout/Layout';
+
 export default function WishPage() {
   return (
-    <div className="w-full">
-      <h3>찜한 상품 페이지</h3>
-    </div>
+    <Layout header title="찜한 상품" canGoBack footer={false}>
+      <div className="w-full">
+        <h3>찜한 상품 페이지</h3>
+      </div>
+    </Layout>
   );
 }
